# docs/openapi.yaml
openapi: 3.0.3
info:
  title: MeetNexus API
  version: 0.1.0
  description: 视频会议业务接口；接口变更必须先更新此文件。
paths:
  /health:
    get:
      operationId: getHealth
      summary: 检查 API 服务是否存活
      description: 仅检查 MeetNexus API 进程是否能够响应，不探测 PostgreSQL、Redis 或 Live777。
      tags:
        - 系统
      responses:
        "200":
          description: API 服务运行正常
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /rooms:
    post:
      operationId: createRoom
      summary: 创建会议
      tags:
        - 房间
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoomRequest"
      responses:
        "201":
          description: 会议创建成功，创建者自动成为主持人
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateRoomResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
  /rooms/{room_id}:
    parameters:
      - $ref: "#/components/parameters/RoomId"
    get:
      operationId: getRoom
      summary: 查询会议及成员在线状态
      tags:
        - 房间
      responses:
        "200":
          description: 查询成功
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomDetailsResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "404":
          $ref: "#/components/responses/RoomNotFound"
  /rooms/{room_id}/members:
    parameters:
      - $ref: "#/components/parameters/RoomId"
    post:
      operationId: joinRoom
      summary: 加入会议
      tags:
        - 房间
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinRoomRequest"
      responses:
        "201":
          description: 加入成功，返回本次会议中的成员编号
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JoinRoomResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "404":
          $ref: "#/components/responses/RoomNotFound"
  /rooms/{room_id}/members/{member_id}:
    parameters:
      - $ref: "#/components/parameters/RoomId"
      - $ref: "#/components/parameters/MemberId"
    delete:
      operationId: leaveRoom
      security:
        - BearerAuth: []
      summary: 离开会议
      tags:
        - 房间
      responses:
        "204":
          description: 已离开会议
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/SessionAuthenticationRequired"
        "403":
          $ref: "#/components/responses/RoomMemberAccessDenied"
        "404":
          $ref: "#/components/responses/MemberNotFound"
  /rooms/{room_id}/members/{member_id}/heartbeat:
    parameters:
      - $ref: "#/components/parameters/RoomId"
      - $ref: "#/components/parameters/MemberId"
    post:
      operationId: refreshRoomMemberPresence
      security:
        - BearerAuth: []
      summary: 刷新成员在线状态
      description: 前端应定期调用；90 秒未刷新时该成员会从 Redis 在线集合过期。
      tags:
        - 房间
      responses:
        "204":
          description: 在线状态已刷新
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/SessionAuthenticationRequired"
        "403":
          $ref: "#/components/responses/RoomMemberAccessDenied"
        "404":
          $ref: "#/components/responses/MemberNotFound"
  /media/whip/{room_id}/{member_id}:
    parameters:
      - $ref: "#/components/parameters/RoomId"
      - $ref: "#/components/parameters/MemberId"
    post:
      operationId: publishRoomMedia
      security:
        - BearerAuth: []
      summary: 发布成员的音视频流
      description: 浏览器提交 WebRTC SDP Offer。服务端校验当前成员身份与发布者一致后，将请求代理至 Live777 WHIP。
      tags:
        - 媒体
      requestBody:
        required: true
        content:
          application/sdp:
            schema:
              type: string
      responses:
        "201":
          description: SDP Answer 已创建
          headers:
            Location:
              $ref: "#/components/headers/MediaSessionLocation"
          content:
            application/sdp:
              schema:
                type: string
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/SessionAuthenticationRequired"
        "403":
          $ref: "#/components/responses/MediaAccessDenied"
        "502":
          $ref: "#/components/responses/MediaServiceUnavailable"
  /media/whep/{room_id}/{member_id}:
    parameters:
      - $ref: "#/components/parameters/RoomId"
      - $ref: "#/components/parameters/MemberId"
    post:
      operationId: subscribeRoomMedia
      security:
        - BearerAuth: []
      summary: 订阅成员的音视频流
      description: 浏览器提交 WebRTC SDP Offer。服务端校验当前成员属于房间后，将请求代理至 Live777 WHEP。
      tags:
        - 媒体
      requestBody:
        required: true
        content:
          application/sdp:
            schema:
              type: string
      responses:
        "201":
          description: SDP Answer 已创建
          headers:
            Location:
              $ref: "#/components/headers/MediaSessionLocation"
          content:
            application/sdp:
              schema:
                type: string
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/SessionAuthenticationRequired"
        "403":
          $ref: "#/components/responses/MediaAccessDenied"
        "502":
          $ref: "#/components/responses/MediaServiceUnavailable"
  /media/sessions/{room_id}/{stream_member_id}/{member_id}/{session_id}:
    parameters:
      - $ref: "#/components/parameters/RoomId"
      - name: stream_member_id
        in: path
        required: true
        description: 被发布媒体的成员编号。
        schema:
          type: string
          format: uuid
      - $ref: "#/components/parameters/MemberId"
      - name: session_id
        in: path
        required: true
        description: Live777 返回的会话编号。
        schema:
          type: string
          minLength: 1
          maxLength: 200
    delete:
      operationId: closeRoomMediaSession
      security:
        - BearerAuth: []
      summary: 关闭媒体会话
      description: 仅允许创建该 WHIP/WHEP 会话的当前成员关闭会话。
      tags:
        - 媒体
      responses:
        "204":
          description: 媒体会话已关闭
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/SessionAuthenticationRequired"
        "403":
          $ref: "#/components/responses/MediaAccessDenied"
        "502":
          $ref: "#/components/responses/MediaServiceUnavailable"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    RoomId:
      name: room_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    MemberId:
      name: member_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  headers:
    RequestId:
      description: 本次请求的关联 ID
      required: true
      schema:
        type: string
        format: uuid
    MediaSessionLocation:
      description: 同源媒体会话资源地址；客户端在关闭 PeerConnection 时使用 DELETE 请求此地址。
      required: true
      schema:
        type: string
        format: uri-reference
  responses:
    SessionAuthenticationRequired:
      description: 未提供、无效或已过期的房间成员会话令牌
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    RoomMemberAccessDenied:
      description: 会话令牌不属于目标房间或成员
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InvalidRequest:
      description: 请求字段或路径参数无效
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    RoomNotFound:
      description: 会议不存在
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    MemberNotFound:
      description: 成员不存在
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    RouteNotFound:
      description: 请求的接口不存在
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    MethodNotAllowed:
      description: 请求方法不受支持
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalServerError:
      description: API 内部错误
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    MediaAccessDenied:
      description: 当前成员无权操作该媒体会话
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    MediaServiceUnavailable:
      description: Live777 媒体服务不可用或媒体协商失败
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    HealthData:
      type: object
      additionalProperties: false
      required:
        - status
        - service
        - version
      properties:
        status:
          type: string
          enum:
            - ok
        service:
          type: string
          enum:
            - meetnexus-api
        version:
          type: string
          example: 0.1.0
    CreateRoomRequest:
      type: object
      additionalProperties: false
      required:
        - title
        - display_name
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 80
          example: 项目每周例会
        display_name:
          type: string
          minLength: 1
          maxLength: 40
          example: 小明
    JoinRoomRequest:
      type: object
      additionalProperties: false
      required:
        - display_name
      properties:
        display_name:
          type: string
          minLength: 1
          maxLength: 40
          example: 小红
    Room:
      type: object
      additionalProperties: false
      required:
        - id
        - title
        - created_at
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        created_at:
          type: string
          format: date-time
    RoomMember:
      type: object
      additionalProperties: false
      required:
        - id
        - display_name
        - role
        - joined_at
        - online
      properties:
        id:
          type: string
          format: uuid
        display_name:
          type: string
        role:
          type: string
          enum:
            - host
            - participant
        joined_at:
          type: string
          format: date-time
        online:
          type: boolean
    RoomDetails:
      type: object
      additionalProperties: false
      required:
        - room
        - members
      properties:
        room:
          $ref: "#/components/schemas/Room"
        members:
          type: array
          items:
            $ref: "#/components/schemas/RoomMember"
    RoomDetailsResponse:
      type: object
      additionalProperties: false
      required:
        - data
        - request_id
      properties:
        data:
          $ref: "#/components/schemas/RoomDetails"
        request_id:
          type: string
          format: uuid
    CreateRoomResponse:
      type: object
      additionalProperties: false
      required:
        - data
        - session_token
        - request_id
      properties:
        data:
          $ref: "#/components/schemas/RoomDetails"
        session_token:
          $ref: "#/components/schemas/SessionToken"
        request_id:
          type: string
          format: uuid
    JoinRoomResponse:
      type: object
      additionalProperties: false
      required:
        - data
        - session_token
        - request_id
      properties:
        data:
          $ref: "#/components/schemas/RoomMember"
        session_token:
          $ref: "#/components/schemas/SessionToken"
        request_id:
          type: string
          format: uuid
    SessionToken:
      type: string
      minLength: 1
      description: 当前房间成员的短期访问令牌；成员接口通过 Authorization Bearer 头传递，禁止写入 URL、日志或持久化存储。
    HealthResponse:
      type: object
      additionalProperties: false
      required:
        - data
        - request_id
      properties:
        data:
          $ref: "#/components/schemas/HealthData"
        request_id:
          type: string
          format: uuid
    ErrorDetail:
      type: object
      additionalProperties: false
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: ROUTE_NOT_FOUND
        message:
          type: string
          example: 请求的接口不存在
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - error
        - request_id
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
        request_id:
          type: string
          format: uuid